# incampo/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Field View Starter API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/plans/{plan_id}/file": {
      "post": {
        "summary": "Upload Plan File",
        "description": "Carica l'immagine della planimetria. Un PDF viene convertito in PNG\n(prima pagina) così tutti i client mostrano un'immagine e basta.\nLe dimensioni in pixel servono ai client per posizionare i pin (x/y relativi).",
        "operationId": "upload_plan_file_plans__plan_id__file_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_plan_file_plans__plan_id__file_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attachments": {
      "post": {
        "summary": "Create Attachment",
        "description": "Crea il record (da web); i byte arrivano dopo con /attachments/{id}/upload.",
        "operationId": "create_attachment_attachments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/attachments/{attachment_id}": {
      "delete": {
        "summary": "Delete Attachment",
        "description": "Soft-delete di foto/firma (chi ha compilato il modulo o creato il task, oppure un manager).",
        "operationId": "delete_attachment_attachments__attachment_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attachments/presign": {
      "post": {
        "summary": "Presign Attachment",
        "description": "L'app chiede dove caricare i byte di un allegato già sincronizzato.\nCon S3 direct: presigned POST sul bucket (tipo e dimensione imposti da S3),\npoi /attachments/{id}/complete. Altrimenti upload multipart sull'API.",
        "operationId": "presign_attachment_attachments_presign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/attachments/{attachment_id}/complete": {
      "post": {
        "summary": "Complete Attachment",
        "description": "Chiude un upload diretto su S3: verifica nel bucket che il file esista, che il\ncontenuto (magic bytes, non il Content-Type dichiarato) sia JPEG/PNG/PDF e\ncoerente con l'estensione, poi valorizza file_url. Idempotente.",
        "operationId": "complete_attachment_attachments__attachment_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attachments/{attachment_id}/upload": {
      "post": {
        "summary": "Upload Attachment",
        "description": "Carica i byte di un allegato. Idempotente: un retry sovrascrive lo stesso file.",
        "operationId": "upload_attachment_attachments__attachment_id__upload_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_attachment_attachments__attachment_id__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/file-links/{key}": {
      "get": {
        "summary": "Get File Link",
        "description": "Come leggere un file (stessi permessi di GET /files): con S3 direct un URL\nfirmato a scadenza da usare così com'è (anche in <img src>, senza JWT);\naltrimenti /files/<key> da chiamare con il JWT.",
        "operationId": "get_file_link_file_links__key__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileLink"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/files/{key}": {
      "get": {
        "summary": "Get File",
        "description": "Serve i file dello storage (filesystem o S3) con il JWT: gli URL nel DB sono sempre /files/<key>.\nSolo ai membri del progetto a cui appartiene il file; file non referenziati → 404.",
        "operationId": "get_file_files__key__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachmentCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Type"
          }
        },
        "type": "object",
        "title": "AttachmentCreate"
      },
      "AttachmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Type"
          },
          "file_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "AttachmentOut"
      },
...
      "Body_upload_attachment_attachments__attachment_id__upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_attachment_attachments__attachment_id__upload_post"
      },
      "Body_upload_plan_file_plans__plan_id__file_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_plan_file_plans__plan_id__file_post"
      },
...
      "FileLink": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "direct": {
            "type": "boolean",
            "title": "Direct"
          },
          "expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "url",
          "direct"
        ],
        "title": "FileLink",
        "description": "URL per leggere un file: firmato e a scadenza (direct, niente JWT) o /files dell'API (con JWT)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanOut": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "file_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Url"
          },
          "width_px": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width Px"
          },
          "height_px": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Px"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "PlanOut"
      },
...
      "PresignRequest": {
        "properties": {
          "attachment_id": {
            "type": "string",
            "title": "Attachment Id"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          }
        },
        "type": "object",
        "required": [
          "attachment_id"
        ],
        "title": "PresignRequest"
      },
      "PresignResponse": {
        "properties": {
          "attachment_id": {
            "type": "string",
            "title": "Attachment Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "max_bytes": {
            "type": "integer",
            "title": "Max Bytes"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Fields",
            "default": {}
          },
          "complete_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complete Url"
          }
        },
        "type": "object",
        "required": [
          "attachment_id",
          "method",
          "upload_url",
          "max_bytes"
        ],
        "title": "PresignResponse",
        "description": "Dove e come caricare i byte di un allegato.\n- filesystem (o S3 senza direct): POST multipart `file` su `upload_url` dell'API, con il JWT;\n- S3 direct: POST multipart su `upload_url` del bucket, SENZA JWT, con `fields` prima\n  del campo `file`; poi POST `complete_url` (API, con JWT) che verifica il file e\n  valorizza file_url."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}